Add back in SemisupervisedMixin for MethylANVI
