+++ src/scvi/model/base/__init__.py
@@ -14,7 +14,7 @@ from ._pyromixin import (
     PyroSviTrainMixin,
 )
 from ._rnamixin import RNASeqMixin
-from ._training_mixin import UnsupervisedTrainingMixin
+from ._training_mixin import SemisupervisedTrainingMixin, UnsupervisedTrainingMixin
 from ._vaemixin import VAEMixin
 
 __all__ = [
@@ -32,4 +32,5 @@ __all__ = [
     "BaseMinifiedModeModelClass",
     "BaseMudataMinifiedModeModelClass",
     "EmbeddingMixin",
+    "SemisupervisedTrainingMixin",
 ]
