+++ src/scvi/model/base/_training_mixin.py
@@ -1,14 +1,28 @@
 from __future__ import annotations
 
+import logging
 from typing import TYPE_CHECKING
 
-from scvi.dataloaders import DataSplitter
+import numpy as np
+import pandas as pd
+import torch
+
+from scvi import REGISTRY_KEYS
+from scvi.data._utils import get_anndata_attribute
+from scvi.dataloaders import DataSplitter, SemiSupervisedDataSplitter
 from scvi.model._utils import get_max_epochs_heuristic, use_distributed_sampler
-from scvi.train import TrainingPlan, TrainRunner
+from scvi.train import SemiSupervisedTrainingPlan, TrainingPlan, TrainRunner
+from scvi.train._callbacks import SubSampleLabels
 from scvi.utils._docstrings import devices_dsp
 
 if TYPE_CHECKING:
+    from collections.abc import Sequence
+
+    from anndata import AnnData
     from lightning import LightningDataModule
+
+
+logger = logging.getLogger(__name__)
 
 
 class UnsupervisedTrainingMixin:
@@ -140,6 +154,197 @@ class UnsupervisedTrainingMixin:
             max_epochs=max_epochs,
             accelerator=accelerator,
             devices=devices,
+            **trainer_kwargs,
+        )
+        return runner()
+
+
+class SemisupervisedTrainingMixin:
+    _training_plan_cls = SemiSupervisedTrainingPlan
+
+    def _set_indices_and_labels(self):
+        """Set indices for labeled and unlabeled cells."""
+        labels_state_registry = self.adata_manager.get_state_registry(REGISTRY_KEYS.LABELS_KEY)
+        self.original_label_key = labels_state_registry.original_key
+        self.unlabeled_category_ = labels_state_registry.unlabeled_category
+
+        labels = get_anndata_attribute(
+            self.adata,
+            self.adata_manager.data_registry.labels.attr_name,
+            self.original_label_key,
+            mod_key=getattr(self.adata_manager.data_registry.labels, "mod_key", None),
+        ).ravel()
+        self._label_mapping = labels_state_registry.categorical_mapping
+
+        # set unlabeled and labeled indices
+        self._unlabeled_indices = np.argwhere(labels == self.unlabeled_category_).ravel()
+        self._labeled_indices = np.argwhere(labels != self.unlabeled_category_).ravel()
+        self._code_to_label = dict(enumerate(self._label_mapping))
+
+    def predict(
+        self,
+        adata: AnnData | None = None,
+        indices: Sequence[int] | None = None,
+        soft: bool = False,
+        batch_size: int | None = None,
+        use_posterior_mean: bool = True,
+    ) -> np.ndarray | pd.DataFrame:
+        """Return cell label predictions.
+
+        Parameters
+        ----------
+        adata
+            AnnData object that has been registered via :meth:`~scvi.model.SCANVI.setup_anndata`.
+        indices
+            Return probabilities for each class label.
+        soft
+            If True, returns per class probabilities
+        batch_size
+            Minibatch size for data loading into model. Defaults to `scvi.settings.batch_size`.
+        use_posterior_mean
+            If ``True``, uses the mean of the posterior distribution to predict celltype
+            labels. Otherwise, uses a sample from the posterior distribution - this
+            means that the predictions will be stochastic.
+        """
+        adata = self._validate_anndata(adata)
+
+        if indices is None:
+            indices = np.arange(adata.n_obs)
+
+        scdl = self._make_data_loader(
+            adata=adata,
+            indices=indices,
+            batch_size=batch_size,
+        )
+        y_pred = []
+        for _, tensors in enumerate(scdl):
+            inference_inputs = self.module._get_inference_input(tensors)
+            batch = inference_inputs.pop(REGISTRY_KEYS.BATCH_KEY)
+
+            cont_key = REGISTRY_KEYS.CONT_COVS_KEY
+            cont_covs = inference_inputs.pop(cont_key, default=None)
+
+            cat_key = REGISTRY_KEYS.CAT_COVS_KEY
+            cat_covs = inference_inputs.pop(cat_key, default=None)
+
+            x = [inference_inputs[key] for key in inference_inputs.keys()]
+
+            pred = self.module.classify(
+                *x,
+                batch_index=batch,
+                cat_covs=cat_covs,
+                cont_covs=cont_covs,
+                use_posterior_mean=use_posterior_mean,
+            )
+            if self.module.classifier.logits:
+                pred = torch.nn.functional.softmax(pred, dim=-1)
+            if not soft:
+                pred = pred.argmax(dim=1)
+            y_pred.append(pred.detach().cpu())
+
+        y_pred = torch.cat(y_pred).numpy()
+        if not soft:
+            predictions = [self._code_to_label[p] for p in y_pred]
+            return np.array(predictions)
+        else:
+            n_labels = len(pred[0])
+            pred = pd.DataFrame(
+                y_pred,
+                columns=self._label_mapping[:n_labels],
+                index=adata.obs_names[indices],
+            )
+            return pred
+
+    @devices_dsp.dedent
+    def train(
+        self,
+        max_epochs: int | None = None,
+        n_samples_per_label: float | None = None,
+        check_val_every_n_epoch: int | None = None,
+        train_size: float = 0.9,
+        validation_size: float | None = None,
+        shuffle_set_split: bool = True,
+        batch_size: int = 128,
+        accelerator: str = "auto",
+        devices: int | list[int] | str = "auto",
+        datasplitter_kwargs: dict | None = None,
+        plan_kwargs: dict | None = None,
+        **trainer_kwargs,
+    ):
+        """Train the model.
+
+        Parameters
+        ----------
+        max_epochs
+            Number of passes through the dataset for semisupervised training.
+        n_samples_per_label
+            Number of subsamples for each label class to sample per epoch. By default, there
+            is no label subsampling.
+        check_val_every_n_epoch
+            Frequency with which metrics are computed on the data for validation set for both
+            the unsupervised and semisupervised trainers. If you'd like a different frequency for
+            the semisupervised trainer, set check_val_every_n_epoch in semisupervised_train_kwargs.
+        train_size
+            Size of training set in the range [0.0, 1.0].
+        validation_size
+            Size of the test set. If `None`, defaults to 1 - `train_size`. If
+            `train_size + validation_size < 1`, the remaining cells belong to a test set.
+        shuffle_set_split
+            Whether to shuffle indices before splitting. If `False`, the val, train,
+            and test set are split in the sequential order of the data according to
+            `validation_size` and `train_size` percentages.
+        batch_size
+            Minibatch size to use during training.
+        %(param_accelerator)s
+        %(param_devices)s
+        datasplitter_kwargs
+            Additional keyword arguments passed into
+            :class:`~scvi.dataloaders.SemiSupervisedDataSplitter`.
+        plan_kwargs
+            Keyword args for :class:`~scvi.train.SemiSupervisedTrainingPlan`. Keyword
+            arguments passed to `train()` will overwrite values present in `plan_kwargs`,
+            when appropriate.
+        **trainer_kwargs
+            Other keyword args for :class:`~scvi.train.Trainer`.
+        """
+        if max_epochs is None:
+            max_epochs = get_max_epochs_heuristic(self.adata.n_obs)
+
+            if self.was_pretrained:
+                max_epochs = int(np.min([10, np.max([2, round(max_epochs / 3.0)])]))
+
+        logger.info(f"Training for {max_epochs} epochs.")
+
+        plan_kwargs = {} if plan_kwargs is None else plan_kwargs
+        datasplitter_kwargs = datasplitter_kwargs or {}
+
+        # if we have labeled cells, we want to subsample labels each epoch
+        sampler_callback = [SubSampleLabels()] if len(self._labeled_indices) != 0 else []
+
+        data_splitter = SemiSupervisedDataSplitter(
+            adata_manager=self.adata_manager,
+            train_size=train_size,
+            validation_size=validation_size,
+            shuffle_set_split=shuffle_set_split,
+            n_samples_per_label=n_samples_per_label,
+            batch_size=batch_size,
+            **datasplitter_kwargs,
+        )
+        training_plan = self._training_plan_cls(self.module, self.n_labels, **plan_kwargs)
+
+        if "callbacks" in trainer_kwargs.keys():
+            trainer_kwargs["callbacks"] + [sampler_callback]
+        else:
+            trainer_kwargs["callbacks"] = sampler_callback
+
+        runner = TrainRunner(
+            self,
+            training_plan=training_plan,
+            data_splitter=data_splitter,
+            max_epochs=max_epochs,
+            accelerator=accelerator,
+            devices=devices,
+            check_val_every_n_epoch=check_val_every_n_epoch,
             **trainer_kwargs,
         )
         return runner()
